Changing coordinator index page to be same as main
